# [Hotfix] Faster get_aggregate_logs_queryset [#OSF-8616]