--- a/osf/models/node.py
+++ b/osf/models/node.py
@@ -779,8 +779,7 @@ class AbstractNode(DirtyFieldsMixin, TypedModel, AddonModelMixin, IdentifierMixi
     def get_aggregate_logs_query(self, auth):
         return (
             (
-                Q(node_id__in=list(Node.objects.get_children(self).can_view(user=auth.user, private_link=auth.private_link).values_list('id', flat=True))) |
-                Q(node_id=self.id)
+                Q(node_id__in=list(Node.objects.get_children(self).can_view(user=auth.user, private_link=auth.private_link).values_list('id', flat=True)) + [self.id])
             ) & Q(should_hide=False)
         )