… osf/migrations/0060_add_nodelog_should_hide_nid_index.py