## Purpose
Prevent deadlock

Before
```
Limit  (cost=0.43..5534.73 rows=6 width=498) (actual time=161265.125..161273.338 rows=6 loops=1)
  ->  Index Scan Backward using osf_nodelog_5fc73231 on osf_nodelog  (cost=0.43..1883507.36 rows=2042 width=498) (actual time=161265.123..161273.333 rows=6 loops=1)
        Filter: ((NOT should_hide) AND ((node_id = ANY ('{<list of ids>}'::integer[])) OR (node_id = <id>)))
        Rows Removed by Filter: 5797099
        SubPlan 2
          ->  Index Scan using osf_abstractnode_pkey on osf_abstractnode u0  (cost=0.42..16.91 rows=1 width=32) (actual time=1.760..1.760 rows=1 loops=6)
                Index Cond: (id = osf_nodelog.node_id)
                SubPlan 1
                  ->  Aggregate  (cost=8.45..8.46 rows=1 width=32) (actual time=0.889..0.889 rows=1 loops=6)
                        ->  Index Scan using osf_guid_content_type_id_cf491d0b_idx on osf_guid  (cost=0.42..8.45 rows=1 width=26) (actual time=0.864..0.864 rows=1 loops=6)
                              Index Cond: ((content_type_id = 30) AND (object_id = u0.id))
        SubPlan 4
          ->  Index Scan using osf_osfuser_pkey on osf_osfuser u0_1  (cost=0.29..16.77 rows=1 width=32) (actual time=0.381..0.381 rows=1 loops=6)
                Index Cond: (id = osf_nodelog.user_id)
                SubPlan 3
                  ->  Aggregate  (cost=8.45..8.46 rows=1 width=32) (actual time=0.183..0.183 rows=1 loops=6)
                        ->  Index Scan using osf_guid_content_type_id_cf491d0b_idx on osf_guid osf_guid_1  (cost=0.42..8.45 rows=1 width=26) (actual time=0.175..0.175 rows=1 loops=6)
                              Index Cond: ((content_type_id = 18) AND (object_id = u0_1.id))
        SubPlan 6
          ->  Index Scan using osf_abstractnode_pkey on osf_abstractnode u0_2  (cost=0.42..16.91 rows=1 width=32) (actual time=0.026..0.027 rows=1 loops=6)
                Index Cond: (id = osf_nodelog.original_node_id)
                SubPlan 5
                  ->  Aggregate  (cost=8.45..8.46 rows=1 width=32) (actual time=0.010..0.010 rows=1 loops=6)
                        ->  Index Scan using osf_guid_content_type_id_cf491d0b_idx on osf_guid osf_guid_2  (cost=0.42..8.45 rows=1 width=26) (actual time=0.003..0.003 rows=1 loops=6)
                              Index Cond: ((content_type_id = 30) AND (object_id = u0_2.id))
Planning time: 38.321 ms
Execution time: 161273.444 ms
```

After
```
Limit  (cost=3871.69..4175.27 rows=6 width=498) (actual time=0.287..0.714 rows=6 loops=1)
  ->  Result  (cost=3871.69..109013.29 rows=2078 width=498) (actual time=0.286..0.711 rows=6 loops=1)
        ->  Sort  (cost=3871.69..3876.88 rows=2078 width=402) (actual time=0.148..0.149 rows=6 loops=1)
              Sort Key: osf_nodelog.date DESC
              Sort Method: top-N heapsort  Memory: 26kB
              ->  Index Scan using osf_nodelog_should_hide_nid on osf_nodelog  (cost=0.43..3834.44 rows=2078 width=402) (actual time=0.016..0.111 rows=123 loops=1)
                    Index Cond: ((should_hide = false) AND (node_id = ANY ('{<list of ids and id>}'::integer[])))
                    Filter: (NOT should_hide)
        SubPlan 2
          ->  Index Scan using osf_abstractnode_pkey on osf_abstractnode u0  (cost=0.42..16.91 rows=1 width=32) (actual time=0.034..0.034 rows=1 loops=6)
                Index Cond: (id = osf_nodelog.node_id)
                SubPlan 1
                  ->  Aggregate  (cost=8.45..8.46 rows=1 width=32) (actual time=0.013..0.013 rows=1 loops=6)
                        ->  Index Scan using osf_guid_content_type_id_cf491d0b_idx on osf_guid  (cost=0.42..8.45 rows=1 width=26) (actual time=0.004..0.004 rows=1 loops=6)
                              Index Cond: ((content_type_id = 30) AND (object_id = u0.id))
        SubPlan 4
          ->  Index Scan using osf_osfuser_pkey on osf_osfuser u0_1  (cost=0.29..16.77 rows=1 width=32) (actual time=0.031..0.031 rows=1 loops=6)
                Index Cond: (id = osf_nodelog.user_id)
                SubPlan 3
                  ->  Aggregate  (cost=8.45..8.46 rows=1 width=32) (actual time=0.008..0.008 rows=1 loops=6)
                        ->  Index Scan using osf_guid_content_type_id_cf491d0b_idx on osf_guid osf_guid_1  (cost=0.42..8.45 rows=1 width=26) (actual time=0.003..0.003 rows=1 loops=6)
                              Index Cond: ((content_type_id = 18) AND (object_id = u0_1.id))
        SubPlan 6
          ->  Index Scan using osf_abstractnode_pkey on osf_abstractnode u0_2  (cost=0.42..16.91 rows=1 width=32) (actual time=0.025..0.025 rows=1 loops=6)
                Index Cond: (id = osf_nodelog.original_node_id)
                SubPlan 5
                  ->  Aggregate  (cost=8.45..8.46 rows=1 width=32) (actual time=0.007..0.007 rows=1 loops=6)
                        ->  Index Scan using osf_guid_content_type_id_cf491d0b_idx on osf_guid osf_guid_2  (cost=0.42..8.45 rows=1 width=26) (actual time=0.002..0.002 rows=1 loops=6)
                              Index Cond: ((content_type_id = 30) AND (object_id = u0_2.id))
Planning time: 23.802 ms
Execution time: 0.799 ms
```



## Changes
* Add index
* Query better

## Side effects
None expected. Requires minimal downtime


## Ticket
[[OSF-8616]](https://openscience.atlassian.net/browse/OSF-8616)